added moderatePrompt tests, added template for prompt tests, moved openAi files into their own folder
