+++ tests/openAi/openAiCompletionsMock.json
@@ -1,5 +1,5 @@
 {
-  "id": "chatcmpl-8wISL0yTfFSpobhpQDZwyXzE0Mmdi",
+  "id": "chatcmpl-XXXXX",
   "object": "chat.completion",
   "created": 1708904661,
   "model": "gpt-3.5-turbo-0125",
